Migrate Auth route to Auth module
