+++ DTOs/UserLogin.cs
@@ -1,0 +1,3 @@
+namespace EventManagement.DTOs;
+
+public record UserLogin(string Email, string Password);
