+++ DTOs/UserRegister.cs
@@ -1,0 +1,3 @@
+namespace EventManagement.DTOs;
+
+public record UserRegister(string Email, string Password);
